# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AiUsageView": {
        "default": null,
        "description": "What this call spent.",
        "properties": {
          "costUsd": {
            "type": "number"
          },
          "durationMs": {
            "format": "int64",
            "type": "integer"
          },
          "inputTokens": {
            "format": "int64",
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "outputTokens": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "BaseScreen": {
        "default": null,
        "description": "The screen that exists in production today, when this is a change request rather than a fresh design.",
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BaseScreenBlock"
            },
            "maxItems": 30,
            "minItems": 0,
            "type": "array"
          },
          "path": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "route": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
      "BaseScreenBlock": {
        "default": null,
        "description": "One block of the live screen being revised.",
        "properties": {
          "kind": {
            "minLength": 1,
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "CanvasBlockView": {
        "default": null,
        "description": "One block of the canvas being edited.",
        "properties": {
          "hidden": {
            "type": "boolean"
          },
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "kind": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "props": {
            "additionalProperties": {},
            "default": "",
            "description": "Catalogue-defined; passed through untouched.",
            "type": "object"
          }
        },
        "required": [
          "id",
          "kind"
        ],
        "type": "object"
      },
      "CanvasEditRequest": {
        "default": null,
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/CanvasBlockView"
            },
            "maxItems": 200,
            "minItems": 0,
            "type": "array"
          },
          "catalog": {
            "default": "",
            "description": "The block catalogue, rendered by the client. This is what tells the model which kinds and props exist.",
            "maxLength": 60000,
            "minLength": 0,
            "type": "string"
          },
          "instruction": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "label": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          },
          "model": {
            "default": null,
            "enum": [
              "opus",
              "sonnet",
              "haiku",
              "opus",
              "sonnet",
              "haiku"
            ],
            "type": "string"
          },
          "projectId": {
            "default": "",
            "description": "Attributes the spend to a project.",
            "type": "string"
          }
        },
        "required": [
          "blocks",
          "catalog",
          "instruction"
        ],
        "type": "object"
      },
      "CanvasEditResponse": {
        "default": null,
        "description": "The model's proposed operations, unvalidated against the catalogue — the client checks them before applying.",
        "properties": {
          "operations": {
            "items": {
              "additionalProperties": {},
              "type": "object"
            },
            "type": "array"
          },
          "reply": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/AiUsageView"
          }
        },
        "type": "object"
      },
      "ChatAttachment": {
        "default": null,
        "properties": {
          "dataBase64": {
            "default": "",
            "description": "Images only — base64 with no data: prefix.",
            "maxLength": 2800000,
            "minLength": 0,
            "type": "string"
          },
          "kind": {
            "enum": [
              "image",
              "text",
              "binary"
            ],
            "type": "string"
          },
          "mediaType": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "text": {
            "default": "",
            "description": "Text files only — their content, read in the browser.",
            "maxLength": 8000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name"
        ],
        "type": "object"
      },
      "ChatRequest": {
        "default": null,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/ChatAttachment"
            },
            "maxItems": 4,
            "minItems": 0,
            "type": "array"
          },
          "context": {
            "maxLength": 200000,
            "minLength": 0,
            "type": "string"
          },
          "effort": {
            "default": null,
            "description": "Thinking depth; ignored by models that do not take it.",
            "enum": [
              "low",
              "medium",
              "high",
              "xhigh",
              "max"
            ],
            "type": "string"
          },
          "folderLabel": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "maxItems": 30,
            "minItems": 0,
            "type": "array"
          },
          "message": {
            "maxLength": 12000,
            "minLength": 0,
            "type": "string"
          },
          "model": {
            "default": null,
            "enum": [
              "opus",
              "sonnet",
              "haiku",
              "opus",
              "sonnet",
              "haiku"
            ],
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "projectName": {
            "maxLength": 160,
            "minLength": 0,
            "type": "string"
          },
          "thinking": {
            "default": false,
            "description": "Ask for a reasoning summary; ignored by models without it.",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
      "ChatTurn": {
        "default": null,
        "properties": {
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "type": "string"
          },
          "text": {
            "maxLength": 6000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "type": "object"
      },
...
      "GenerateScreensRequest": {
        "default": null,
        "properties": {
          "baseScreen": {
            "$ref": "#/components/schemas/BaseScreen"
          },
          "catalog": {
            "maxLength": 60000,
            "minLength": 0,
            "type": "string"
          },
          "customer": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "maxScreens": {
            "format": "int32",
            "maximum": 8,
            "minimum": 1,
            "type": "integer"
          },
          "model": {
            "default": null,
            "enum": [
              "opus",
              "sonnet",
              "haiku",
              "opus",
              "sonnet",
              "haiku"
            ],
            "type": "string"
          },
          "notes": {
            "maxLength": 12000,
            "minLength": 20,
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "referenceNames": {
            "default": "",
            "description": "Names of attachments with no extractable text, so the model knows they exist.",
            "items": {
              "maxLength": 200,
              "minLength": 0,
              "type": "string"
            },
            "maxItems": 40,
            "minItems": 0,
            "type": "array"
          },
          "references": {
            "default": "",
            "description": "Merged text of the meeting's reference files.",
            "maxLength": 16000,
            "minLength": 0,
            "type": "string"
          },
          "sessionTitle": {
            "maxLength": 160,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "catalog",
          "notes"
        ],
        "type": "object"
      },
      "GenerateScreensResponse": {
        "default": null,
        "properties": {
          "reply": {
            "type": "string"
          },
          "screens": {
            "items": {
              "$ref": "#/components/schemas/GeneratedScreen"
            },
            "type": "array"
          },
          "skipped": {
            "default": "",
            "description": "Screens the model proposed that this API could not read at all.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/AiUsageView"
          }
        },
        "type": "object"
      },
      "GeneratedBlock": {
        "default": null,
        "properties": {
          "kind": {
            "minLength": 1,
            "type": "string"
          },
          "props": {
            "additionalProperties": {},
            "default": "",
            "description": "Catalogue-defined; passed through untouched.",
            "type": "object"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "GeneratedScreen": {
        "default": null,
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/GeneratedBlock"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "route": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WorkspaceStatePatch": {
        "default": null,
        "description": "A batch of writes, applied together. The browser batches its changes and sends them as one request rather than one per key.",
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceStateWrite"
            },
            "maxItems": 500,
            "minItems": 0,
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "type": "object"
      },
      "WorkspaceStateResult": {
        "default": null,
        "properties": {
          "deleted": {
            "format": "int32",
            "type": "integer"
          },
          "rejected": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "written": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "WorkspaceStateSnapshot": {
        "default": null,
        "description": "Everything the caller can see, as one map. Read once when the workspace loads, because the code that consumes it reads synchronously.",
        "properties": {
          "entries": {
            "additionalProperties": {
              "type": "string"
            },
            "default": "",
            "description": "Key to the exact string that was stored.",
            "type": "object"
          },
          "personalKeys": {
            "default": "",
            "description": "Keys that belong to this viewer rather than the team.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "WorkspaceStateWrite": {
        "default": null,
        "description": "One key to write, or to delete when `value` is null.",
        "properties": {
          "key": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          },
          "value": {
            "default": "",
            "description": "The string to store. Null deletes the key.",
            "maxLength": 8000000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Projects, meetings, generated screens, and the requests they become.",
    "license": {
      "name": "Proprietary"
    },
    "title": "WE-ADK API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai/canvas": {
      "post": {
        "description": "The operations come back as the model wrote them. They are validated against the block catalogue on the client, which owns it.",
        "operationId": "canvas",
        "parameters": [
          {
            "in": "header",
            "name": "X-Anthropic-Api-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanvasEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CanvasEditResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Turn an instruction into canvas operations.",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/ai/chat": {
      "post": {
        "operationId": "chat",
        "parameters": [
          {
            "description": "Spend your own quota instead of the server's.",
            "in": "header",
            "name": "X-Anthropic-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "additionalProperties": {
                    "default": ""
                  },
                  "default": "",
                  "type": "string"
                }
              }
            },
            "description": "A stream of envelopes: `stream_event` per token, then a closing `result`."
          }
        },
        "summary": "One turn of the folder chat, streamed as newline-delimited JSON.",
        "tags": [
          "AI"
        ]
      }
    },
...
    "/api/ai/generate": {
      "post": {
        "operationId": "generate",
        "parameters": [
          {
            "in": "header",
            "name": "X-Anthropic-Api-Key",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateScreensRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateScreensResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Propose the concept screens a set of meeting notes implies.",
        "tags": [
          "AI"
        ]
      }
    },
...
    "/api/state": {
      "get": {
        "description": "One request rather than one per key: the code that consumes this reads synchronously from inside render paths, so it hydrates once at load.",
        "operationId": "snapshot",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStateSnapshot"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Everything the caller can see, in one read.",
        "tags": [
          "Workspace state"
        ]
      },
      "put": {
        "operationId": "apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceStatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStateResult"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Apply a batch of writes and deletions.",
        "tags": [
          "Workspace state"
        ]
      }
    },
    "/api/state/import": {
      "post": {
        "description": "The one-time handover from a browser that still holds state. Fills gaps and never overwrites, so whichever browser connects first cannot flatten work done from another.",
        "operationId": "importMissing",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceStatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceStateResult"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Seed only the keys that do not exist yet.",
        "tags": [
          "Workspace state"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}